still buiding the table/forms

+++ db/define_schema.rb
@@ -19,7 +19,7 @@ DB.define_column("swimmer_infos", "first_name", "string")
 DB.define_column("swimmer_infos", "team_id", "integer")
 DB.define_column("swimmer_infos", "age", "integer")
 DB.define_column("swimmer_infos", "gender_id", "integer")
-DB.define_column("swimmer_infos", "height", "integer")
+DB.define_column("swimmer_infos", "height", "float")
 DB.define_column("swimmer_infos", "weight", "integer")
 
 DB.define_table("genders")
